Stop screwing with META-INF for versioning info
Fixes GH-8

This seems less to do with which version of Java being run, and more to
do with the overall environment its running in. For example, I was
constantly seeing it try and load Gnome Accessibility stuff from a
manifest that I can't even find in our jar. That combined with the
overall 'fat-jar' approach we're taking, it's much simpler and more
cosnsistent to just load it from our own file.