Make discord better handle whitespace-only messages
Prior to this change an IRC message that was just a space would trigger
an exception upon being sent to Discord as Discord doesn't allow that.
In order to remain true to IRC, we will broadcast a zero-width space in
its place, which Discord seems to be just fine with.