Add initial improvements to bridging Discord replies
Closes GH-48

Prior to this change, Discord replies would be bridged without any
context. Now, there will be additional context bridged along with the
"reply" message.

The default is to limit the length of this context to 90 characters, so
as to avoid filling the screen. This is user configurable under the IRC
config key `discord-reply-context-limit`.

To completely disable this new behavior, and not bridge any context
whatsoever, you can set the value to 0.

20:37 <B> Reply to "Z752: I really enjoy potatoes, I think they're quite amazing"
20:37 <B> <Z750> What? Heresy! Everyone knows tacos are superior!