Minor fixes for Discord reply bridging
Translate the referenced message as well as the parent when bridging.
Prior to this the referenced message would use the incorrect formatting
codes (and not have any mutators applied to it).

Remove newlines from the context message. As part of the "don't fill up
the screen" rationale from the original commit, we don't want newlines
making their way into the context messages. They are now replaced with
spaces.

Also corrects a typo.