Add channel mentions support
Fixes GH-53

Also invert the default condition of requiring whitespace separation for the
replace util. Most mentions in discord do not require it so there's no
reason to have it on by default. This change also corrects several other
mentions that did require that separation that do not in discord proper.