Initial support for Discord stickers
Lottie-based JSON stickers are NOT supported at this time. We would need
to find a public web player somewhere to hook into to support this. In
my testing, not all libraries support these properly anyway. Some are
messing up the colors. It may be easier to just self-host this
ourselves.

Image-based PNG/APNG stickers appear to be working using the
media.discordapp.com endpoint. Whether that changes in the future, who
knows.

Adds feature requested in GH-57