Add support for JSON/Lottie stickers
These Lottie stickers are basically vector animations, so we must pass
them along to a web viewer for handling. I am currently self-hosting the
one I found to work the best (closest to Discord's).

Unfortunately, this viewer requires JS to be enabled to run. Nothing
much we can do about that. Well, nothing _sane_ we can do about that
anyway.