Add colored nicknames to more places in IRC
This commit fixes a bug we had in the development branch (not any
releases) where anti-ping and colored nick would conflict with one
another.

As a side effect, we now use color in more places when that option is
enabled, such as the configurable command forward messages and the
discord reply context messages.