Allow overriding the build suffix
Prior to this commit, the build suffix (displayed to users at startup)
was always the top of tree git short hash. This is nice but it falls
down if someone is pulling a release from a tarball, zip file, or other
archive that does not include the .git folder or if the builder does not
have git installed.

As of this commit, the project will now build even in those cases
(displaying a "0" suffix). This change also allows the builder to
specify the suffix for packaging or other purposes via the "suffix"
property. E.g. `./gradlew build -P suffix="build35"`

I have no immediate plans to change my use of the git hash in releases
on GitHub, but hopefully this allows others additional flexibility.

Fixes #82