Print markdown symbols in IRC for code blocks, bold, etc.
Ideally, this would be configured in the settings, but for now it is
unconditional.

In many IRC clients, including goguma/gamja/senpai (all the clients
with good support for soju bouncers like <https://chat.sr.ht>), the
formatting Dis4IRC previously used for this is invisible. For them,
preserving the backticks, starts, etc, is very helpful.

Partially fixes #81