Refactor bridge message types into disparate classes
Refactor our message type classes so that only the relevant data is
exposed on each type of message. E.g. a message that originates from IRC
no longer has access to an attachments field, as obviously IRC messages
do not have attachments.

This ends up being quite a large refactor and affects a substantial
amount of the codebase. Testing was done to ensure basic functionality
however it would not surprise me if a new bug or two was introduced by
this change. However, it's also at least possible that this clarity has
fixed an old issue or two (though I can't think of one off the top of
my head).