Ensure message has destination before processing
We should not be running mutators or doing any processing on messages
before we determine they have a destination. There may be messages
coming in on other channels/sources that are not intended to be bridged.

Previously, we would still run mutators on these messages. They were not
bridged, but they were eating unnecessary resources. This appears to
have been introduced in the recent refactor.