Add warning message for users not on Java 21
Due to a variety of factors, including library dependencies looking to
update their own minimum version of Java (e.g. outside our control), we
should warn users that future versions will likely require a newer
version of Java.

There is no set timeline for requiring Java 21 on Dis4IRC's part.
If/when we need to require the newer version, we will defer the change
until the next feature release. E.g. 1.X.0, where X indicates the
feature version.

Most recent Linux distributions and *nix package managers already offer
Java 21 (and some default to it) so this is not expected to be a major
problem for most - however, we should be nice and try and warn those who
are not on it.